move metrics logic to `bee-node`
