+++ bee-node/bee-node/src/config.rs
@@ -17,10 +17,7 @@ use bee_ledger::workers::{
 };
 #[cfg(feature = "dashboard")]
 use bee_plugin_dashboard::config::{DashboardConfig, DashboardConfigBuilder};
-use bee_protocol::workers::{
-    config::{ProtocolConfig, ProtocolConfigBuilder},
-    MetricsConfig, MetricsConfigBuilder,
-};
+use bee_protocol::workers::config::{ProtocolConfig, ProtocolConfigBuilder};
 use bee_rest_api::endpoints::config::{RestApiConfig, RestApiConfigBuilder};
 use bee_tangle::config::{TangleConfig, TangleConfigBuilder};
 use fern_logger::{LoggerConfig, LoggerConfigBuilder, LOGGER_STDOUT_NAME};
@@ -28,7 +25,13 @@ use serde::Deserialize;
 
 #[cfg(feature = "trace")]
 use crate::trace::{TraceConfig, TraceConfigBuilder};
-use crate::{cli::ClArgs, storage::NodeStorageBackend, util, BECH32_HRP_DEFAULT, NETWORK_NAME_DEFAULT};
+use crate::{
+    cli::ClArgs,
+    storage::NodeStorageBackend,
+    util,
+    workers::{MetricsConfig, MetricsConfigBuilder},
+    BECH32_HRP_DEFAULT, NETWORK_NAME_DEFAULT,
+};
 
 pub(crate) const ALIAS_DEFAULT: &str = "bee";
