+++ bee-node/bee-node/src/workers.rs
@@ -1,0 +1,93 @@
+// Copyright 2020-2022 IOTA Stiftung
+// SPDX-License-Identifier: Apache-2.0
+
+use std::{convert::Infallible, net::SocketAddr, time::Duration};
+
+use async_trait::async_trait;
+use bee_metrics::{metrics::ProcessMetrics, Registry};
+use bee_runtime::{node::Node, shutdown_stream::ShutdownStream, worker::Worker};
+use futures::StreamExt;
+use log::info;
+use serde::Deserialize;
+use tokio::time::interval;
+use tokio_stream::wrappers::IntervalStream;
+
+const PROCESS_METRICS_UPDATE_INTERVAL: Duration = Duration::from_secs(5);
+
+const DEFAULT_BIND_ADDRESS: &str = "0.0.0.0:3030";
+
+#[derive(Default, Deserialize, PartialEq)]
+#[must_use]
+pub struct MetricsConfigBuilder {
+    #[serde(skip)]
+    pid: Option<u32>,
+    #[serde(alias = "bindAddress")]
+    bind_address: Option<SocketAddr>,
+}
+
+impl MetricsConfigBuilder {
+    pub fn with_pid(mut self, pid: u32) -> Self {
+        self.pid = Some(pid);
+        self
+    }
+
+    pub fn finish(self) -> MetricsConfig {
+        MetricsConfig {
+            pid: self.pid.unwrap(),
+            bind_address: self
+                .bind_address
+                .unwrap_or_else(|| DEFAULT_BIND_ADDRESS.parse().unwrap()),
+        }
+    }
+}
+
+#[derive(Clone)]
+pub struct MetricsConfig {
+    pid: u32,
+    bind_address: SocketAddr,
+}
+
+pub struct MetricsRegistryWorker {}
+
+#[async_trait]
+impl<N: Node> Worker<N> for MetricsRegistryWorker {
+    type Config = MetricsConfig;
+    type Error = Infallible;
+
+    #[cfg_attr(feature = "trace", trace_tools::observe)]
+    async fn start(node: &mut N, config: Self::Config) -> Result<Self, Self::Error> {
+        node.register_resource(Registry::default());
+        let registry = node.resource::<Registry>();
+
+        let process_metrics = ProcessMetrics::new(config.pid);
+        let (mem_metric, cpu_metric) = process_metrics.metrics();
+
+        registry.register("bee_memory_usage", "Memory currently allocated by the node", mem_metric);
+
+        registry.register("bee_cpu_usage", "CPU pecentage currently used by the node", cpu_metric);
+
+        node.spawn::<Self, _, _>(|shutdown| async move {
+            let mut ticker =
+                ShutdownStream::new(shutdown, IntervalStream::new(interval(PROCESS_METRICS_UPDATE_INTERVAL)));
+
+            while ticker.next().await.is_some() {
+                process_metrics.update().await;
+            }
+        });
+
+        node.spawn::<Self, _, _>(|shutdown| async move {
+            info!("Running prometheus client.");
+
+            let serve = bee_metrics::serve_metrics(config.bind_address, registry);
+
+            tokio::select! {
+                _ = shutdown => (),
+                _ = serve => (),
+            };
+
+            info!("Stopped prometheus client.");
+        });
+
+        Ok(Self {})
+    }
+}
