+++ bee-protocol/src/workers/mod.rs
@@ -22,10 +22,7 @@ mod status;
 
 use bee_autopeering::event::EventRx as AutopeeringEventRx;
 use bee_gossip::NetworkEventReceiver as NetworkEventRx;
-use bee_runtime::{
-    node::{Node, NodeBuilder},
-    worker::Worker,
-};
+use bee_runtime::node::{Node, NodeBuilder};
 
 use self::peer::PeerManagerConfig;
 pub(crate) use self::{
@@ -49,7 +46,7 @@ pub(crate) use self::{
 };
 pub use self::{
     message::{MessageSubmitterError, MessageSubmitterWorker, MessageSubmitterWorkerEvent},
-    metrics::{MetricsConfig, MetricsConfigBuilder, MetricsWorker},
+    metrics::MetricsWorker,
     peer::{PeerManager, PeerManagerResWorker},
     requester::{request_message, MessageRequesterWorker, RequestedMessages, RequestedMilestones},
 };
@@ -60,13 +57,12 @@ pub fn init<N: Node>(
     network_events: NetworkEventRx,
     autopeering_events: Option<AutopeeringEventRx>,
     node_builder: N::Builder,
-    metrics_config: <MetricsWorker as Worker<N>>::Config,
 ) -> N::Builder
 where
     N::Backend: storage::StorageBackend,
 {
     node_builder
-        .with_worker_cfg::<MetricsWorker>(metrics_config)
+        .with_worker::<MetricsWorker>()
         .with_worker::<PeerManagerResWorker>()
         .with_worker_cfg::<PeerManagerWorker>(PeerManagerConfig {
             network_rx: network_events,
